Corrections...

+++ WebTier-Image-recognition-Iaas/src/main/java/com/aws/cse546/aws_Iaas_image_recognition/webTier/services/ImageRecognitionWebTierService.java
@@ -1,11 +1,15 @@
 package com.aws.cse546.aws_Iaas_image_recognition.webTier.services;
 
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+import org.apache.commons.codec.binary.Base64;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -33,18 +37,20 @@ public class ImageRecognitionWebTierService implements Runnable {
 		this.outputMap = outputMap;
 	}
 
-	
 	public String[] getOutputFromResponseQueue(String imageUrl) {
 		System.out.println("ImageURL: " + imageUrl);
 		while (true) {
 			try {
+				System.out.println("Trying....!");
 				if (this.outputMap.containsKey(imageUrl)) {
+					System.out.println("got in" + imageUrl);
+					logger.info("Got the image with URL - {}", imageUrl);
 					String output = this.outputMap.get(imageUrl);
 					this.outputMap.remove(imageUrl);
 					return new String[] { this.formatImageUrl(imageUrl), output };
 				} else {
 					try {
-						Thread.sleep(8000);
+						Thread.sleep(2000);
 					} catch (Exception e) {
 						e.printStackTrace();
 					}
@@ -52,7 +58,7 @@ public class ImageRecognitionWebTierService implements Runnable {
 			} catch (Exception e) {
 				System.out.println("Some Error while getting outPut from HashMap");
 				try {
-					Thread.sleep(8000);
+					Thread.sleep(1000);
 				} catch (Exception o) {
 					o.printStackTrace();
 				}
@@ -83,9 +89,16 @@ public class ImageRecognitionWebTierService implements Runnable {
 					try {
 						for (Message msg : msgList) {
 							String[] classificationResult = null;
+							logger.info(msg.getBody());
 							classificationResult = msg.getBody().split(ProjectConstants.INPUT_OUTPUT_SEPARATOR);
-							outputMap.put(classificationResult[0], classificationResult[1]);
-							awsService.deleteMessage(msg, ProjectConstants.OUTPUT_QUEUE);
+							if(classificationResult.length > 1) {
+								outputMap.put(classificationResult[0], classificationResult[1]);
+								logger.info("Received Message from response queue: {}", classificationResult[0]+ " - "+ classificationResult[1]);
+								awsService.deleteMessage(msg, ProjectConstants.OUTPUT_QUEUE);
+							}else {
+								logger.error("Message is not proper");
+							}
+								
 						}
 					} catch (Exception w) {
 						logger.info("Error in putting message from queue to map");
@@ -122,6 +135,25 @@ public class ImageRecognitionWebTierService implements Runnable {
 		} catch (Exception e) {
 			return null;
 		}
+	}
+	
+	public String getBase64OutofImage(File file) {
+		String encodedfile = null;
+        try {
+            try (FileInputStream fileInputStreamReader = new FileInputStream(file)) {
+				byte[] bytes = new byte[(int)file.length()];
+				fileInputStreamReader.read(bytes);
+				encodedfile = new String(Base64.encodeBase64(bytes), "UTF-8");
+			}
+        } catch (FileNotFoundException e) {
+            logger.info("file not found while encoding the image");
+            e.printStackTrace();
+        } catch (IOException e) {
+            logger.info("Some error occurred while encoding!");
+            e.printStackTrace();
+        }
+
+        return encodedfile;
 	}
 
 }
